+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added `pybricksdev oad info` command.
+
 ## [1.0.0-alpha.50] - 2024-07-01
 
 ### Changed